Move include from last line
